# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "NFT Music",
            "url": "https://nft.threenext.com",
            "email": "seki@threenext.com"
        }
    },
    "paths": {
...
        "/genres": {
            "get": {
                "description": "ジャンルマスターの情報をリストで取得する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ジャンルマスター"
                ],
                "summary": "ジャンルマスターの情報をリストで取得する",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ports.GenreMasterOutput"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "post": {
                "description": "ジャンルマスターの情報を1件作成する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ジャンルマスター"
                ],
                "summary": "ジャンルマスターの情報を1件作成する",
                "parameters": [
                    {
                        "description": "ジャンルマスター",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.GenreMasterInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.GenreMasterOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
        "/genres/{id}": {
            "get": {
                "description": "ジャンルマスターの情報を出力し",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ジャンルマスター"
                ],
                "summary": "ジャンルマスターから特定の１レコードを抽出する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ジャンルマスターID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.GenreMasterOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "put": {
                "description": "ジャンルマスターの情報を1件修正する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ジャンルマスター"
                ],
                "summary": "ジャンルマスターの情報を1件修正する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ジャンルマスターID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ジャンルマスター",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.GenreMasterInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.GenreMasterOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "delete": {
                "description": "ジャンルマスターの情報を1件削除する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ジャンルマスター"
                ],
                "summary": "ジャンルマスターの情報を1件削除する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ジャンルマスターID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
        "/ipfs": {
            "post": {
                "description": "分散型ストレージIPFSに画像を登録する",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPFS"
                ],
                "summary": "IPFSノードにイメージデータを登録",
                "parameters": [
                    {
                        "type": "file",
                        "description": "this is a test file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ウォレットアドレス",
                        "name": "wallet",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.IpfsOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
        "/ipfs/meta": {
            "post": {
                "description": "分散型ストレージIPFSにJSONデータを登録する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPFS"
                ],
                "summary": "IPFSノードにJSONデータを登録",
                "parameters": [
                    {
                        "description": "MetaJSON",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.IpfsMetaInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.IpfsOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ports.ErrorResponseObject": {
            "type": "object",
            "properties": {
                "error_type": {
                    "type": "string",
                    "example": "リクエストの構文が不正やページが見つからない等のタイプ別"
                },
                "message": {
                    "type": "string",
                    "example": "原因となるのエラーメッセージ"
                },
                "status_code": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "ports.GenreMasterInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "エレクトロニック・ダンス・ミュージック"
                }
            }
        },
        "ports.GenreMasterOutput": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-11-04T20:51:26Z"
                },
                "id": {
                    "type": "string",
                    "example": "01932573-3f06-78c6-9743-b495cfcc2167"
                },
                "name": {
                    "type": "string",
                    "example": "エレクトロニック・ダンス・ミュージック"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-11-04T20:51:26Z"
                }
            }
        },
        "ports.IpfsMetaInput": {
            "type": "object",
            "required": [
                "audio_cid",
                "description",
                "file_type",
                "image_cid",
                "name",
                "video_cid"
            ],
            "properties": {
                "audio_cid": {
                    "type": "string",
                    "example": "QmPYVinZ3fmWSetwc9FWMF9b93hfKjWiHtygXfVY8exwEM"
                },
                "description": {
                    "type": "string",
                    "example": "良いNFTです"
                },
                "file_type": {
                    "type": "string",
                    "example": "audio"
                },
                "image_cid": {
                    "type": "string",
                    "example": "QmW9qWKbZneDijFPcHdbn41fQiHRwownM4U6avLsCHfMJS"
                },
                "insentive": {
                    "type": "integer",
                    "example": 10
                },
                "name": {
                    "type": "string",
                    "example": "GoodNFT"
                },
                "video_cid": {
                    "type": "string",
                    "example": "QmQfESxcop5u9zjvK5hJyX8oKJJZCsEYVh4kTU64Wzmxqz"
                }
            }
        },
        "ports.IpfsOutput": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}